issue/#186 refactor: 스탬프 찍기 API 별도 분리 및 해당 패스 수정
